fix: fix Header with children
