+++ src/components/ui/header/header.module.scss
@@ -1,43 +1,11 @@
 .headerWrapper {
-  width: 1280px;
-  height: 60px;
-  padding: 0 136px;
   display: flex;
-  justify-content: space-between;
   align-items: center;
+  justify-content: space-between;
+
+  width: 80rem;
+  height: 3.75rem;
+  padding: 0 8.5rem;
+
   background: var(--color-dark-700);
-
-  .logo {
-    width: 156px;
-    height: 36px;
-  }
-
-  .withAvatar {
-    display: flex;
-    align-items: center;
-    gap: 14px;
-    position: relative;
-
-    .avatar {
-      cursor: pointer;
-      width: 36px;
-      height: 36px;
-      border-radius: 50%;
-    }
-
-    .name {
-      border-bottom: 1px dashed var(--color-light-100);
-    }
-    .headerMenu {
-      position: absolute;
-      top: 50px;
-      left: -100px;
-
-      /*& img {
-        max-width: 36px;
-        max-height: 36px;
-        border-radius: 50%;
-      }*/
-    }
-  }
 }
