--- src/components/ui/header/header.stories.ts
@@ -1,25 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { Header } from './'
-
-const meta = {
-  title: 'Components/Header',
-  component: Header,
-  tags: ['autodocs'],
-  argTypes: {},
-} satisfies Meta<typeof Header>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const HeaderWithButton: Story = {
-  args: {
-    isSignedIn: false,
-  },
-}
-
-export const HeaderWithDropDownMenu: Story = {
-  args: {
-    isSignedIn: true,
-  },
-}
